+++ src/Views/BucketPanel/types.ts
@@ -1,0 +1,6 @@
+import { ReadableRecord } from "reduct-js/lib/cjs/Record";
+
+export type IndexedReadableRecord = {
+  record: ReadableRecord;
+  tableIndex: number;
+};